refactor(neo-tree): separate out plugin specific portions of `neo-tree` configuration
