+++ lua/astronvim/plugins/neo-tree.lua
@@ -201,33 +201,6 @@ return {
         vim.opt_local.foldcolumn = "0"
       end,
     })
-
-    if astro.is_available "telescope.nvim" then
-      opts.commands.find_in_dir = function(state)
-        local node = state.tree:get_node()
-        local path = node.type == "file" and node:get_parent_id() or node:get_id()
-        require("telescope.builtin").find_files { cwd = path }
-      end
-      opts.window.mappings.F = "find_in_dir"
-    end
-
-    if astro.is_available "toggleterm.nvim" then
-      local function toggleterm_in_direction(state, direction)
-        local node = state.tree:get_node()
-        local path = node.type == "file" and node:get_parent_id() or node:get_id()
-        require("toggleterm.terminal").Terminal:new({ dir = path, direction = direction }):toggle()
-      end
-      local prefix = "T"
-      ---@diagnostic disable-next-line: assign-type-mismatch
-      opts.window.mappings[prefix] =
-        { "show_help", nowait = false, config = { title = "New Terminal", prefix_key = prefix } }
-      for suffix, direction in pairs { f = "float", h = "horizontal", v = "vertical" } do
-        local command = "toggleterm_" .. direction
-        opts.commands[command] = function(state) toggleterm_in_direction(state, direction) end
-        opts.window.mappings[prefix .. suffix] = command
-      end
-    end
-
     return opts
   end,
 }
