+++ lua/astronvim/plugins/telescope.lua
@@ -91,6 +91,20 @@ return {
         }
       end,
     },
+    {
+      "nvim-neo-tree/neo-tree.nvim",
+      optional = true,
+      opts = {
+        commands = {
+          find_in_dir = function(state)
+            local node = state.tree:get_node()
+            local path = node.type == "file" and node:get_parent_id() or node:get_id()
+            require("telescope.builtin").find_files { cwd = path }
+          end,
+        },
+        window = { mappings = { F = "find_in_dir" } },
+      },
+    },
   },
   dependencies = {
     "nvim-treesitter/nvim-treesitter",
